requirements.txt
numpy
pandas
scikit-learn
xgboost
joblib
pyyaml
matplotlib
shap
streamlit    # optional for demo
python-dotenv
